remove gui_job_submittal component; fix the branch selection re-rendering bug by changing job_submittal to a pure component
